DFR-1460: Remove 'Trial bundle' option on 'upload case file' event
