+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/UploadContestedCaseDocumentsAboutToSubmitHandler.java
@@ -14,6 +14,7 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.ContestedUploadedDocument;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.ContestedUploadedDocumentData;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.FeatureToggleService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.casedocuments.CaseDocumentHandler;
 
 import java.util.ArrayList;
@@ -21,6 +22,7 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.CONTESTED_UPLOADED_DOCUMENTS;
 
@@ -30,6 +32,12 @@ import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigCo
 public class UploadContestedCaseDocumentsAboutToSubmitHandler
     implements CallbackHandler<Map<String, Object>> {
 
+    public static final String TRIAL_BUNDLE_SELECTED_ERROR =
+        "To upload a hearing bundle please use the Manage hearing "
+            + "bundles event which can be found on the drop-down list on the home page";
+    public static final String TRIAL_BUNDLE_TYPE = "Trial Bundle";
+
+    private final FeatureToggleService featureToggleService;
     private final List<CaseDocumentHandler> caseDocumentHandlers;
     private final ObjectMapper objectMapper;
     private final UploadedDocumentHelper uploadedDocumentHelper;
@@ -55,10 +63,9 @@ public class UploadContestedCaseDocumentsAboutToSubmitHandler
                 ContestedUploadedDocument::getCaseDocumentUploadDateTime, Comparator.nullsLast(
                     Comparator.reverseOrder()))));
         caseData.put(CONTESTED_UPLOADED_DOCUMENTS, uploadedDocuments);
-        return GenericAboutToStartOrSubmitCallbackResponse
-            .<Map<String, Object>>builder()
-            .data(caseData)
-            .build();
+        GenericAboutToStartOrSubmitCallbackResponse<Map<String, Object>> response = getCallBackResponse(caseData);
+        setWarningsAndErrors(caseData, response);
+        return response;
     }
 
     private List<ContestedUploadedDocumentData> getDocumentCollection(Map<String, Object> caseData) {
@@ -69,6 +76,39 @@ public class UploadContestedCaseDocumentsAboutToSubmitHandler
             });
 
         return Optional.ofNullable(contestedUploadDocuments).orElse(new ArrayList<>());
+    }
+
+    private GenericAboutToStartOrSubmitCallbackResponse<Map<String, Object>> getCallBackResponse(Map<String, Object> caseData) {
+        return GenericAboutToStartOrSubmitCallbackResponse.<Map<String, Object>>builder()
+            .data(caseData)
+            .errors(new ArrayList<>())
+            .warnings(new ArrayList<>())
+            .build();
+    }
+
+    private void setWarningsAndErrors(Map<String, Object> caseData, GenericAboutToStartOrSubmitCallbackResponse<Map<String, Object>> response) {
+        if (featureToggleService.isManageBundleEnabled()
+            && isTrialBundleSelectedInAnyUploadedFile(caseData)) {
+            response.getErrors().add(TRIAL_BUNDLE_SELECTED_ERROR);
+        }
+    }
+
+    private boolean isTrialBundleSelectedInAnyUploadedFile(Map<String, Object> caseData) {
+        return !getTrialBundleUploadedList(getDocumentCollection(caseData)).isEmpty();
+    }
+
+    private List<ContestedUploadedDocumentData> getTrialBundleUploadedList(List<ContestedUploadedDocumentData> uploadedDocuments) {
+
+        return uploadedDocuments.stream()
+            .filter(d -> isTrialBundle(d.getUploadedCaseDocument()))
+            .collect(Collectors.toList());
+    }
+
+    private boolean isTrialBundle(ContestedUploadedDocument uploadedCaseDocument) {
+        return Optional.ofNullable(uploadedCaseDocument)
+            .map(ContestedUploadedDocument::getCaseDocumentType)
+            .filter(type -> type.equals(TRIAL_BUNDLE_TYPE))
+            .isPresent();
     }
 
 }
